Have `extend_syntax` take a BNF extension directly, not a function on `SynEnv`s.
For typechecking reasons, `extend_syntax` needs to be able to reach into
each `Syntax` (i.e. a RHS of the BNF) and get the type environment it
produces. So this means that general `SynEnv`-to-`SynEnv` functions
won't work, unless they ... also returned a reified version of the
resulting type environment (or enough stuff to compute it)? Ew!

Anyways, I think there's rarely going to be a need to programmatically
decide which NTs to add to the syntax, so this should be fine. This
undoes the work for #23, though. Now it's more of a WONTFIX, I guess.

Also, fix a couple of bugs in core_macro_forms.rs.