Macro expansion works now!
Decided to put macro implementations in `Form.eval`, instead of the
`Ast` itself, because it made more sense. (Note that having a separate
field from `eval` would be neater, but I couldn't bear to add yet
another field.)

This turns out to be a simpler way of handling the macro implementation
hygeine problem, too. Closes #25.