Change the standard extension for Unseemly from `.≉` to `.unseemly`
(Not that it's really enforced in any way at this point)

Using the logo of a programming language as a file extension was cute,
and I was thinking that it was okay to do something weird like this
because very little code ought to be written in plain Unseemly, since,
you know, it doesn't even have comments.

But the thing is that it *looks* user-hostile, so, even if it provably
isn't a technical obstacle to anyone, it has an unfriendly feel. Too
bad, I liked it.