A couple of hacks to make code inside macro implementations work.
I think there needs to be a `--` (`QuoteLess`) in `expand_syntax` or
macro definitions in order to make the phases work right, but I'm not
sure why I had to make the imports phaseless AND put the `--` inside
them.

Also, allow `--`ing into "negative" phases (actually, I think that I got
it backwards and those are positive phases in the Scheme sense).

Also, apply a hack so that walks "know" their default literalness (issue